refactor: change file directory and apply CLS prevention
